Don't error out if ${go_pkg_dir} already exists
If it already exists, the world is OK and there's no need to panic.

+++ hack/lib/golang.sh
@@ -270,7 +270,9 @@ kube::golang::create_gopath_tree() {
   rm -f "${go_pkg_dir}"
 
   # TODO: This symlink should be relative.
-  ln -s "${KUBE_ROOT}" "${go_pkg_dir}"
+  if [[ ! -e "${go_pkg_dir}" ]]; then
+    ln -s "${KUBE_ROOT}" "${go_pkg_dir}"
+  fi
 }
 
 # Ensure the godep tool exists and is a viable version.
